rename former example dir from v6
