+++ examples-v6/assets/icons.ts
