+++ examples-v6/assets/person.png
