+++ examples-v6/components/index.ts
