+++ examples-v6/fixed-nodes/data.ts
