+++ examples-v6/fixed-nodes/index.html
