+++ examples-v6/hierarchy/index.ts
