+++ examples-v6/landing/data.ts
