+++ examples-v6/pixi-icons/index.html
