+++ examples-v6/pixi-icons/index.ts
@@ -2,7 +2,7 @@ import Stats from 'stats.js'
 import * as Force from '../../src/layout/force'
 import * as Fisheye from '../../src/layout/fisheye'
 import * as Cluster from '../../src/layout/cluster'
-import * as Graph from '../../src/'
+import * as Graph from '../../src'
 import * as WebGL from '../../src/renderers/webgl'
 import { company } from '../assets/icons'
 import person from '../assets/person.png'
