+++ examples-v6/pixi-perf/index.html
