+++ examples-v6/pixi-react/index.html
