+++ examples-v6/pixi-react/index.ts
