+++ examples-v6/pixi-render-perf/index.ts
