+++ examples-v6/pixi/index.ts
@@ -7,7 +7,7 @@ import * as Download from '../../src/bindings/native/download'
 import * as Cluster from '../../src/layout/cluster'
 import * as WebGL from '../../src/renderers/webgl'
 import * as Png from '../../src/renderers/image'
-import * as Graph from '../../src/'
+import * as Graph from '../../src'
 
 export const stats = new Stats()
 stats.showPanel(0) // 0: fps, 1: ms, 2: mb, 3+: custom
