+++ examples-v6/radial/index.ts
@@ -1,6 +1,6 @@
 import Stats from 'stats.js'
 import * as Radial from '../../src/layout/radial'
-import * as Graph from '../../src/'
+import * as Graph from '../../src'
 import * as WebGL from '../../src/renderers/webgl'
 import graphData from '../../data/tmp-data'
