+++ examples-v6/tidy-tree/data.ts
