+++ examples-v6/tsconfig.json
